Move some functions to where I think they make more sense
